% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_plans.R
\name{load_plans}
\alias{load_plans}
\title{load_plans}
\usage{
load_plans()
}
\value{
returns a data object with the plans references in the rct function
}
\description{
load_plans
}
\examples{
load_plans()
}
