% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct.R
\name{rct}
\alias{rct}
\title{rct function}
\usage{
rct(cases, type = "square", border = "#1239DC", xlim = c(0, 1),
  ylim = c(0, 1), fill = NULL, xlab = NULL, ylab = "",
  lab.cex = 1, seed = NULL, plot.new = TRUE, label = FALSE,
  lab.col = "#1239DC", draw.plot = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{
returns a risk characterization theatre plot
}
\description{
rct function
}
